Simplify VRL_STATES by incorporating additional methods into the enum
Previously, the globals.py used a separate dictionary named "G_VEHICLE_STATUS_DICT" to hold the names of the individual vehicle states.
The code is simplified using tuples for the values of enum in the VRL_STATES. The value method is overriden to return the state value (first element of the tuple) and the method "display_name" returns the second element of the tuple.