Introduce a new class StationaryProcess to represent the task performed at a PlanStop
A PlanStop can consists of tasks performed on stationary vehicles (e.g. charging, boarding, etc). As a PlanStop is mostly used for making planned scheduled, the exact details of these tasks can be delegated to StationaryProcess class with standard interfaces to the PlanStop.
Additionally, the same StationaryProcess instance (copied from the PlanStop instance) is used to perform the actual stationary task