DEBUG: call the time trigger of the charging operator while finishing the remaining tasks at the end of simulation
Some of the simulations with the public charging operator was raising assertion error while finishing the remaining tasks at the end of simulation. This espeially occured while using pre-calculated bookings by private vehicles.

The reason is that the time trigger of the charging operator was not being called in the method "record_remaining_assignments". The time trigger removes any expired booking, e.g., the virtual bookings by the private vehicles that have crossed the "end time".