FEAT: Make the router NetworkTTMatrix able to dynamically load time-dependent travel time matrices
The router checks if the provided string for the parameter "nw_dynamic_f" in the simulation configuration is a file or a folder. If it is a folder, then the subfolder names represent simulation time. The numpy file "tt_matrix.npy" in the subfolders are then loaded accordign to the current simulation time.