feat: only use the distance criterion for route when the travel time criterion raises error
Sometimes, using both distance and time critria together raises error as the error margin is different in both criteria.
Therefore, the 1-1 route finding on the NetworkTTMatrix router now first uses the time criteria alone. If a route is not found using it, then it uses time and distance criteria together.